Refactor CLI structure to modular command organization
- Restructured the CLI by moving command implementations into separate modules under the `commands` directory.
- Introduced new command files for `fetch`, `reply`, `resolve`, and `schema`, enhancing maintainability and readability.
- Updated the main CLI entry point to register these commands, streamlining the command handling process.
- Improved error handling and validation across commands, ensuring consistent user feedback and experience.
- Adjusted tests to reflect the new command structure, ensuring all functionalities are covered and operational.
