+++ src/toady/commands/__init__.py
@@ -1,0 +1,1 @@
+"""Command modules for Toady CLI."""
